cleaning up backend calls -> spot_calls.py, rm sample_db
